+++ CHANGELOG.md
@@ -15,6 +15,8 @@ viewer;
 
 ## Changed
 
+- concord-server: keep the original values of `readonly` form fields
+on submit;
 - concord-server: throttle the AD/LDAP user group caching;
 - concord-console: fixed the log toolbar's flickering issue.